+++ spec/ig_markets/cli/commands/transactions_command_spec.rb
@@ -21,14 +21,14 @@ END
                                                                                    ).to_stdout
   end
 
-  it 'prints transactions from a number of days and a start time' do
+  it 'prints transactions using a from and to time' do
     from = Time.new 2015, 2, 15, 6, 0, 0, '-04:00'
     to = Time.new 2015, 2, 18, 6, 0, 0, '-04:00'
 
     expect(dealing_platform.account).to receive(:transactions).with(from: from, to: to).and_return([])
 
     expect do
-      cli(days: 3, from: '2015-02-15T06:00:00-04:00', interest: true).transactions
+      cli(from: '2015-02-15T06:00:00-04:00', to: '2015-02-18T06:00:00-04:00').transactions
     end.to output("#{IGMarkets::CLI::Tables::TransactionsTable.new []}\n").to_stdout
   end